further refactoring of duplicated code into AbstractGroupingAction
